+++ packages/dothtml/src/reactivity/binding.ts
@@ -2,6 +2,7 @@ import { IBinding, IWatcher } from "dothtml-interfaces";
 import Signal from "./signal";
 
 export default class Binding<T = any, Td = T> implements IBinding<T, Td> {
+	_isBinding = true;
 	_source: IWatcher<T>;
 
 	_transform: {