+++ packages/interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # dothtml-interfaces
 
+## 6.0.5
+
+### Patch Changes
+
+- Polymorphic mounting.
+
 ## 6.0.4
 
 ### Patch Changes